requirements.txt
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.11.0
tqdm>=4.65.0
scikit-learn>=1.3.0
openpyxl>=3.1.0
argparse
